% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{applyROIOnImage}
\alias{applyROIOnImage}
\title{Apply binary mask (ROI) on the image to retain only the region of
interest}
\usage{
applyROIOnImage(img, roi)
}
\arguments{
\item{img}{An \code{EBImage::Image} object to apply ROI on}
}
\value{
Returns the image with only the region of interest and the
background removed
}
\description{
Apply binary mask (ROI) on the image to retain only the region of
interest
}
