% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{applyTransformOnMask}
\alias{applyTransformOnMask}
\title{Transform a mask using a \code{SimpleITK}'s transformation object}
\usage{
applyTransformOnMask(mask, tf)
}
\arguments{
\item{mask}{Mask to be transformed}

\item{tf}{Transformation object from \code{SimpleITK} (generally a
\code{CompositeTransform} object)}
}
\value{
Returns a binary matrix after transformation
}
\description{
\code{applyTransformOnMask()} applies the transformation from
\code{SimpleITK}. Resampling is done using \code{SimpleITK}.
}
