% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convertToImage}
\alias{convertToImage}
\title{Convert feature matrices with ROI into rasterized matrices with
features as channels}
\usage{
convertToImage(img, roi = NULL, attrs = NULL)
}
\arguments{
\item{img}{A \code{Rtsne::Rtsne} or a matrix with ROI}

\item{roi}{A logical rasterized matrix; each element denotes if the
pixel is present}

\item{attrs}{A list of attributes; at a minimum, it should contain
nX and nY, the dimensions of img}
}
\value{
Returns an \code{EBImage::Image} in \code{Colormode} if \code{img} has more
than 1 column else \code{Grayscale}
}
\description{
Convert feature matrices with ROI into rasterized matrices with
features as channels
}
