% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coregister.R
\name{coregister}
\alias{coregister}
\title{Co-registers MS imaging data with microscopic optical images}
\usage{
coregister(
  mse,
  opt,
  mse_roi = NULL,
  opt_roi = NULL,
  SSC = TRUE,
  mz_list = NULL,
  spatial_scale = 1,
  BPPARAM = SerialParam(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mse}{The Mass Spectrometry (MS) imaging data.}

\item{opt}{The optical image of class \code{EBImage::Image}.}

\item{mse_roi}{A binary mask identifying the  tissue or region-of-interest
on the MS image.}

\item{opt_roi}{A binary mask identifying the tissue or region-of-interest
on the optical image.}

\item{SSC}{A logical flag (default = TRUE); whether to perform Spatial
Shrunken Centroids method from \code{Cardinal} for feature selection.}

\item{mz_list}{A vector of mass / charge (mz) values (defualt = NULL); SSC
will be skipped and the \code{mz_list} will be passed for dimensionality
reduction.}

\item{spatial_scale}{Factor with respect to MS image's spatial dimensions
(default = 1); co-registration will be performed at this scale.}

\item{BPPARAM}{\code{BiocParallelParam} specification (default = \code{SerialParam()});
See documentation for \code{bplapply}.}

\item{verbose}{A logical flag (default=FALSE); whether updates shoule be
printed.}

\item{...}{Keyword arguments passed to \code{SimpleITK}'s registration methods.}
}
\value{
A list of items
Summarized MS image at its original spatial resolution
Processed optical image at its original spatial resolution
Summarized MS image at scaled spatial resolution
Processed optical image at its scaled spatial resolution
Registration object from SimpleITK
Composite transform of the registration
}
\description{
\code{coregister()} processes and summarizes MS imaging data and optical images
to make it feasible to apply \code{SimpleITK}'s family of registration methods.
This method uses spatial shrunken centroids method from \code{Cardinal} to
filtering out unimportant features.
}
