% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{cropToEdgesAndPadZeros}
\alias{cropToEdgesAndPadZeros}
\title{Approximately match positions of tissues (binary masks) in two images}
\usage{
cropToEdgesAndPadZeros(ref_mask, target_mask, target_img = NULL)
}
\arguments{
\item{ref_mask}{A reference binary mask showing position of the tissue}

\item{target_mask}{A binary mask of the tissue to be spatially matched
to that of \code{ref_mask}}

\item{target_img}{An optional image (\code{EBImage::Image}) which, if
provided, will be spatially matched to the tissue in \code{ref_mask}}
}
\value{
Returns a list of two items: \code{target_mask} and \code{target_img} in
which the tissue positions are now approximately in the same area
as that of in \code{ref_mask}
}
\description{
\code{cropToEdgesAndPadZeros()} builds target image and its corresponding
(spatially matching) target mask to approximately match the
spatial position the tissue (ROI) on the optical image with that of
the MSI image.
}
