% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{diceCoefFromMasks}
\alias{diceCoefFromMasks}
\title{Evaluate dice coefficient between two masks / ROIs}
\usage{
diceCoefFromMasks(mask1, mask2)
}
\arguments{
\item{mask1}{A binary matrix consisting of a mask / ROI of an image}

\item{mask2}{A binary matrix consisting of a mask / ROI of a
corresponding image}
}
\value{
Returns the dice coefficient (a measure of agreement or
alignment) between the mask pair.
}
\description{
\code{diceCoefFromMasks()} evaluates the dice coefficient between two
corresponding masks or ROIs. Masks are considered binary
matrices with the tissue labeled \code{TRUE} and background labeled
\code{FALSE}.
}
