% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{displacementFieldJacobianDeterminant}
\alias{displacementFieldJacobianDeterminant}
\title{Evaluate the jacobian determinant from a displacement
field}
\usage{
displacementFieldJacobianDeterminant(displ_field = NULL)
}
\arguments{
\item{displ_field}{Displacement field (vector) of a
transformation}
}
\value{
Returns a matrix of the same size as that of the
\code{displ_field}. Values of the matrix are the jacobian
deterimant at that position
}
\description{
\code{displacementFieldJacobianDeterminant()} computes the jacobian
determinant from a displacement field. Displacement field
should be (X, Y, 2) shape, where X and Y are spatial
dimensions of the images.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Jacobian determinant
    = 1: no transformation 
    > 1: expansion 
    < 1: contraction 
    < 0: folding
}\if{html}{\out{</div>}}
}
