% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{intensityMatrix2D}
\alias{intensityMatrix2D}
\title{Build the intensity matrix from MS image as a feature matrix i.e.,
the matrix's rows are features and columns are pixels of MS
image}
\usage{
intensityMatrix2D(mse, byrow = FALSE)
}
\arguments{
\item{mse}{An MS imaging dataset}

\item{byrow}{A logical flag (default = FALSE); whether the pixels
should be ordered as row or column major}
}
\value{
Returns a 2D matrix of intensities
}
\description{
Build the intensity matrix from MS image as a feature matrix i.e.,
the matrix's rows are features and columns are pixels of MS
image
}
