% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{jacobianDeterminantFromTransform}
\alias{jacobianDeterminantFromTransform}
\title{Evaluate the jacobian determinant of a transform}
\usage{
jacobianDeterminantFromTransform(
  tf,
  dim = NULL,
  ref_img = NULL,
  simpleITK_for_det = TRUE
)
}
\arguments{
\item{tf}{\code{SimpleITK} transform object (generally a
\code{CompositeTransform} object)}

\item{dim}{Spatial dimensions of the reference image (either
this parameter or \code{ref_img} needs to be provided)}

\item{ref_img}{A reference image (generally the moving image)
(either \code{dim} or this parameter needs to be provided)}
}
\value{
Returns a matrix of the same size as \code{dim} or that of
the \code{ref_img}. Values of the matrix are the jacobian
deterimant at that position
}
\description{
\code{jacobianDeterminantFromTransform()} computes the jacobian
determinant from a \code{SimpleITK} transform object.
Jacobian determinant informs the transformation quality
such as image folding.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Jacobian determinant
    = 1: no transformation 
    > 1: expansion 
    < 1: contraction 
    < 0: folding
}\if{html}{\out{</div>}}
}
