% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{multiDrawROI}
\alias{multiDrawROI}
\title{Select multiple regions-of-interest on optical image}
\usage{
multiDrawROI(img)
}
\arguments{
\item{img}{An \code{EBImage::Image} object to select ROI on}
}
\value{
Returns the ROI as a logical matrix of the same spatial dimensions
as \code{img}
}
\description{
Select multiple regions-of-interest on optical image
}
