% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{multiSelectROI}
\alias{multiSelectROI}
\title{Select multiple regions-of-interest on Mass Spectrometry image}
\usage{
multiSelectROI(mse, mz, rasterize = TRUE, ...)
}
\arguments{
\item{mse}{The imaging dataset to select ROI}

\item{mz}{An example ion image with mass / charge (mz) to display; passed
over to \code{Cardinal::image()} function}

\item{rasterize}{A logical flag (default = TRUE); whether to rasterize the
ROI (affects the cases only where MS image is not rectangular).}

\item{...}{Passed to \code{Cardinal::selectROI()}}
}
\value{
Returns the ROI as a logical matrix if \code{rasterize} is TRUE or a
logical vector if \code{rasterize} is FALSE
}
\description{
Select multiple regions-of-interest on Mass Spectrometry image
}
