% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{normalizedCrossCorr}
\alias{normalizedCrossCorr}
\title{Evaluate the normalized cross correlation metric}
\usage{
normalizedCrossCorr(img1, img2)
}
\arguments{
\item{img1}{An intensity array of an image}

\item{img2}{An intensity array of the corresponding image}
}
\value{
Returns a value of the normalized cross correlation
}
\description{
\code{normalizedCrossCorr()} computes the normalized cross correlation
between a pair of images. The images have to be identical in
dimensions.
}
