% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{resizeMask}
\alias{resizeMask}
\title{Resize a mask to given dimensions}
\usage{
resizeMask(ref, target)
}
\arguments{
\item{ref}{Dimensions (a vector) to which the mask (\code{target}) is
resized}

\item{target}{Binary mask (matrix) to be resized}
}
\value{
Returns a binary matrix resized to the given dimensions
}
\description{
\code{resizeMask()} resizes \code{target} mask to a given dimensions (\code{ref}).
Resizing is done using bilinear interpolation. \code{target} mask
should be a binary matrix.
}
