% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI.R
\name{multiDrawROI}
\alias{multiDrawROI}
\title{Select multiple regions-of-interest (ROIs) on optical image}
\usage{
multiDrawROI(img)
}
\arguments{
\item{img}{An \code{EBImage::Image} object to select ROI on}
}
\value{
Returns the ROI as a logical matrix of the same spatial
dimensions as \code{img}
}
\description{
\code{multiDrawROI()} extends \code{drawROIOnImage()} to allow selection of
multiple disconnected tissues in the optical image.
}
